% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_axis.R
\name{component_axis}
\alias{component_axis}
\title{lattice arrow style axis}
\usage{
component_axis(
  ticks,
  labels = TRUE,
  origin = 0,
  tck = 0.4,
  angle = 0,
  type = c("xaxis", "yaxis")[1],
  title = TRUE
)
}
\arguments{
\item{ticks}{numeric vector}

\item{labels}{boolean or character vector with the same length as \code{ticks}. If
\code{labels = FALSE}, no lables will be displayed.}

\item{origin}{begin point of axis}

\item{tck}{tck length in the unit of lines}

\item{angle}{axis text angle}

\item{type}{one of "xaxis" and "yaxis"}
}
\description{
lattice arrow style axis
}
