% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_polygon.R
\name{draw_polygon}
\alias{draw_polygon}
\title{draw horizontal and vertical polygons}
\usage{
draw_polygon(
  vals,
  x = NULL,
  type = "horizontal",
  length.out = 10000,
  alpha = 0.6,
  ...
)
}
\arguments{
\item{vals}{numeric vector}

\item{x}{The corresponding x position of \code{vals}}

\item{type}{one of "horizontal" or "vertical"}

\item{length.out}{the length of interpolated \code{vals} and \code{x}}

\item{alpha}{the alpha of polygon's fill color}
}
\description{
draw horizontal and vertical polygons
}
\examples{
set.seed(1)
y <- rnorm(10)
x <- seq_along(y)
}
