% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dvolcano}
\alias{dvolcano}
\title{Topographic Information on Auckland's Maunga Whau dvolcano}
\format{
An object of class \code{data.frame} with 5307 rows and 3 columns.
}
\source{
Digitized from a topographic map by Ross Ihaka. These data should
not be regarded as accurate.
}
\usage{
dvolcano
}
\description{
Maunga Whau (Mt Eden) is one of about 50 dvolcanos in the Auckland volcanic
field. This data set gives topographic information for Maunga Whau on a 10m by 10m grid.
}
\keyword{internal}
