% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{get_color}
\alias{get_color}
\alias{get_color_symmetry}
\title{Get and interpolate colors}
\usage{
get_color(col, n = NULL, show = FALSE)

get_color_symmetry(x = "amwg256", at, origin = 0, dark = TRUE)
}
\description{
Get and interpolate colors
}
