% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{grid.draw2}
\alias{grid.draw2}
\title{convert plot into grob and draw grob}
\usage{
grid.draw2(p)
}
\arguments{
\item{p}{one of the supported figure}
}
\description{
convert plot into grob and draw grob
}
\keyword{internal}
