% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesc.R
\name{imagesc}
\alias{imagesc}
\title{imagesc}
\usage{
imagesc(
  arr,
  x = NULL,
  y = NULL,
  sp = TRUE,
  cols,
  brks,
  xlab = "",
  ylab = "",
  sp.layout = NULL,
  colorkey = NULL,
  labeller = labeller_num(),
  par.settings = NULL,
  ...
)
}
\description{
imagesc
}
