% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticeMap-package.R
\name{init_lattice}
\alias{init_lattice}
\title{initialize the new lattice functions in arrangeGrob}
\usage{
init_lattice()
}
\description{
initialize the new lattice functions in arrangeGrob
}
\keyword{internal}
