% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice.layers-package.R
\docType{package}
\name{lattice.layers-package}
\alias{lattice.layers}
\alias{lattice.layers-package}
\title{lattice.layers: Lattice Layers Functions}
\description{
Based on lattice, this package extends Lattice Panels by `panel.annotation`, and provides several high-level function for preoducing figure.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
