% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticeGrob-package.R
\docType{package}
\name{latticeGrob-package}
\alias{latticeGrob}
\alias{latticeGrob-package}
\title{latticeGrob: Extend Lattice Panels by Grob Objects}
\description{
Based on lattice, this package extends Lattice Panels by `panel.annotation`,
    and provides several high-level function for preoducing figure.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
