% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_annotation.R
\name{layer_annotation}
\alias{layer_annotation}
\title{layer_annotation}
\usage{
layer_annotation(
  grob,
  bbox = c(0.5, 1, 0, 1),
  unit = "npc",
  xscale = FALSE,
  yscale = FALSE,
  clip = "on",
  ...
)
}
\arguments{
\item{grob}{grob object}

\item{bbox}{The region to plot \code{grob}, \verb{[xmin, xmax, ymin, ymax]} in the unit
of npc.}

\item{unit}{A character vector specifying the units for the corresponding numeric values.}

\item{xscale}{A numeric vector of length two indicating the minimum and
    maximum on the x-scale.  The limits may not be identical. }

\item{yscale}{A numeric vector of length two indicating the minimum
    and maximum on the y-scale. The limits may not be identical. }

\item{clip}{One of \code{"on"}, \code{"inherit"}, or
    \code{"off"}, indicating whether to
    clip to the extent of this viewport, inherit the clipping region
    from the parent viewport, or turn clipping off altogether.
    For back-compatibility, a logical value of \code{TRUE} corresponds
    to \code{"on"} and \code{FALSE} corresponds to \code{"inherit"}.

    May also be a grob (or a gTree) that describes a clipping path.}

\item{...}{ignored}
}
\description{
layer_annotation
}
