% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_barchart.R
\name{layer_barchart}
\alias{layer_barchart}
\alias{lay_hisFreq}
\alias{panel.histFreq}
\title{layer of frequency histogram}
\usage{
layer_barchart(
  x = 0.05,
  y = 0.15,
  width = 0.25,
  height = 0.4,
  yticks = seq(0, 0.3, 0.1),
  box.width = 0.8,
  digit = 2,
  title = TRUE,
  xlabels = TRUE,
  ...
)

lay_hisFreq(
  x = 0.05,
  y = 0.15,
  width = 0.25,
  height = 0.4,
  yticks = seq(0, 0.3, 0.1),
  box.width = 0.8,
  digit = 2,
  title = TRUE,
  xlabels = TRUE,
  ...
)

panel.histFreq(
  z,
  subscripts,
  at,
  col.regions,
  x = 0.05,
  y = 0.15,
  width = 0.25,
  height = 0.4,
  yticks = NULL,
  box.width = 0.8,
  digit = 2,
  unit = "npc",
  title = TRUE,
  xlabels = TRUE,
  border = "transparent",
  yscale = NULL,
  ntick = NULL,
  w = NULL,
  ...
)
}
\arguments{
\item{x}{ Variables defining the plot. }

\item{y}{ Variables defining the plot. }

\item{yticks}{yaxis ticks position, in the range of \verb{[0, 100]}}

\item{box.width}{ Thickness of bars in absolute units; overrides
    \code{box.ratio}.  Useful for specifying thickness when the
    categorical variable is not a factor, as use of \code{box.ratio}
    alone cannot achieve a thickness greater than 1. }

\item{digit}{the digit in the labels}

\item{...}{other parameters to \code{\link[=panel.latFreq]{panel.latFreq()}}}

\item{z}{ Variables defining the plot. }

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }

\item{at}{
    Numeric vector giving breakpoints along the range of
    \code{z}.  See \code{\link[lattice]{levelplot}} for details.
  }

\item{col.regions}{
    A vector of colors, or a function to produce a vecor of colors, to
    be used if \code{region=TRUE}.  Each interval defined by \code{at}
    is assigned a color, so the number of colors actually used is one
    less than the length of \code{at}.  See \code{\link[lattice]{level.colors}}
    for details on how the color assignment is done.
  }

\item{border}{ Border color for rectangles used when \code{region=TRUE}. }

\item{yscale}{frequency range, \verb{[0, 100]}}

\item{ntick}{Integer, number of ticks}

\item{w}{weights for percentage, with the same length as \code{z[subscripts]}}
}
\description{
layer of frequency histogram
}
\examples{
\dontrun{
panel.barchartFreq(z, subscripts, bbox = c(0.05, 0.5, 0, 0.5), unit = "npc")
}
}
