% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_signDist.R
\name{layer_signDist}
\alias{layer_signDist}
\alias{panel.signDist}
\title{layer_signDist}
\usage{
layer_signDist(col = "black", lty = 1, lwd = 1, density = 1, angle = 45, ...)

panel.signDist(
  mask,
  SpatialPixel,
  col = "black",
  lty = 1,
  lwd = 1,
  density = 1,
  angle = 45,
  ...
)
}
\arguments{
\item{density}{A small (big) value corresponds to a low (high) contour density.}
}
\description{
layer_signDist
}
