% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_statistic.R
\name{layer_statistic}
\alias{layer_statistic}
\title{layer of spatial statistic label}
\usage{
layer_statistic(
  x = 0.1,
  y = 0.9,
  digit = 2,
  include.sd = TRUE,
  name = "u",
  unit = "",
  FUN = weightedMedian,
  hjust = 0,
  vjust = 1,
  ...
)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{name}{ A character identifier. }

\item{FUN}{one of \code{\link[matrixStats:weightedMedian]{matrixStats::weightedMedian()}}, \code{\link[matrixStats:weightedMean]{matrixStats::weightedMean()}}}

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{...}{other parameters to \code{\link[=grid.text]{grid.text()}}}
}
\description{
layer of spatial statistic label
}
