% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_title.R
\name{layer_title}
\alias{layer_title}
\title{layer_title}
\usage{
layer_title(x = 0, y = 1, cex = 1.4, labels = NULL, hjust = 0, vjust = 1, ...)
}
\arguments{
\item{x}{A numeric vector or unit object specifying x-values.}

\item{y}{A numeric vector or unit object specifying y-values.}

\item{hjust}{A numeric vector specifying horizontal justification.
    If specified, overrides the \code{just} setting.}

\item{vjust}{A numeric vector specifying vertical justification.
    If specified, overrides the \code{just} setting.}

\item{...}{other parameters to \code{\link[=grid.text]{grid.text()}}}
}
\description{
layer_title
}
