% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.annotation.R
\name{panel.annotation}
\alias{panel.annotation}
\title{panel.annotation}
\usage{
panel.annotation(
  grob,
  bbox = c(0.5, 1, 0, 1),
  xscale = FALSE,
  yscale = FALSE,
  clip = "off",
  ...
)
}
\arguments{
\item{grob}{grob object}

\item{bbox}{The region to plot \code{grob}, \verb{[xmin, xmax, ymin, ymax]} in the unit
of npc.}

\item{xscale, yscale}{}

\item{...}{ignored}
}
\description{
panel.annotation
}
