% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchartFreq.R
\name{panel.barchartFreq}
\alias{panel.barchartFreq}
\title{Panel of frequency barchart}
\usage{
panel.barchartFreq(
  z,
  subscripts,
  bbox = c(0.05, 0.5, 0, 0.5),
  unit = "npc",
  at,
  col.regions,
  box.width = 0.8,
  border = "transparent",
  yscale = NULL,
  yticks = NULL,
  ntick = NULL,
  w = NULL,
  digit = 2,
  ...
)
}
\arguments{
\item{z}{ Variables defining the plot. }

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }

\item{bbox}{The region to plot \code{grob}, \verb{[xmin, xmax, ymin, ymax]} in the unit
of npc.}

\item{unit}{A character vector specifying the units for the corresponding numeric values.}

\item{at}{
    Numeric vector giving breakpoints along the range of
    \code{z}.  See \code{\link[lattice]{levelplot}} for details.
  }

\item{col.regions}{
    A vector of colors, or a function to produce a vecor of colors, to
    be used if \code{region=TRUE}.  Each interval defined by \code{at}
    is assigned a color, so the number of colors actually used is one
    less than the length of \code{at}.  See \code{\link[lattice]{level.colors}}
    for details on how the color assignment is done.
  }

\item{box.width}{ Thickness of bars in absolute units; overrides
    \code{box.ratio}.  Useful for specifying thickness when the
    categorical variable is not a factor, as use of \code{box.ratio}
    alone cannot achieve a thickness greater than 1. }

\item{border}{ Border color for rectangles used when \code{region=TRUE}. }

\item{yscale}{frequency range, \verb{[0, 100]}}

\item{yticks}{yaxis ticks position, in the range of \verb{[0, 100]}}

\item{ntick}{Integer, number of ticks}

\item{w}{weights for percentage, with the same length as \code{z[subscripts]}}

\item{digit}{the digit in the labels}

\item{...}{ Extra parameters. }
}
\description{
Panel of frequency barchart
}
\examples{
\dontrun{
panel.barchartFreq(z, subscripts, bbox = c(0.05, 0.5, 0, 0.5), unit = "npc")
}
}
