% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.horizontalFreq.R
\name{panel.horizontalFreq}
\alias{panel.horizontalFreq}
\alias{panel.verticalFreq}
\title{Panel of horizontal and vertical polygons}
\usage{
panel.horizontalFreq(
  x,
  y,
  z,
  subscripts,
  ...,
  tck = 0.02,
  xlab = "",
  ylab = ""
)

panel.verticalFreq(
  x,
  y,
  z,
  subscripts,
  ...,
  tck = 0.02,
  xlab = "",
  ylab = "",
  xlabels = NULL,
  ylabels = NULL,
  digit = 1
)
}
\arguments{
\item{x}{ Variables defining the plot. }

\item{y}{ Variables defining the plot. }

\item{z}{ Variables defining the plot. }

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }

\item{...}{ Extra parameters. }

\item{tck}{The length of tick marks as a fraction of the smaller of the width
or height of the plotting region. If tck >= 0.5 it is interpreted as a fraction
of the relevant side, so if tck = 1 grid lines are drawn. The default setting
(tck = NA) is to use tcl = -0.5.}

\item{xlab, ylab}{the title of xaxis and yaxis.}
}
\description{
Panel of horizontal and vertical polygons
}
