% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.sign.R
\name{panel.sign}
\alias{panel.sign}
\title{Panel of percentage of negative (significant negative\%) and positive\% (significant positive\%)}
\usage{
panel.sign(z = NULL, mask = NULL, xpos = 0.1, ypos = 0.9, ...)
}
\arguments{
\item{z}{numeric vector}

\item{mask}{boolean vector with the same length as z, indicating whether
corresponding z value is significant.}

\item{xpos, ypos}{The x and y position of positive and negative percentage label.}

\item{...}{ignored}
}
\description{
Panel of percentage of negative (significant negative\%) and positive\% (significant positive\%)
}
