% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.spatial.R
\name{panel.spatial}
\alias{panel.spatial}
\alias{panel.spatialBasic}
\title{panel.spatial}
\usage{
panel.spatial(
  x,
  y,
  z,
  subscripts,
  contour = FALSE,
  class = NULL,
  interpolate = TRUE,
  list.mask = NULL,
  SpatialPixel = NULL,
  ...,
  sp.layout
)

panel.spatialBasic(
  x,
  y,
  z,
  subscripts,
  class,
  contour,
  interpolate,
  ...,
  sp.layout = NULL
)
}
\arguments{
\item{x, y, z}{ Variables defining the plot. }

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }

\item{contour}{
    A logical flag, specifying whether contour lines should be drawn.
  }

\item{interpolate}{ logical, passed to \code{\link{grid.raster}}. }

\item{list.mask}{NULL or list of masks. \code{mask} is a boolean vector, \code{TRUE}
indicates significant.}

\item{SpatialPixel}{corresponding SpatialPixel object of \code{df}

If \code{mask} present in \code{df}, \code{df.mask} will be ignored.}

\item{...}{ Extra parameters. }

\item{sp.layout}{NULL or list; see notes below}

\item{par}{position of title, \code{list(title = list(x, y))}}
}
\description{
panel.spatial
}
