% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lattice.R
\name{+.trellis}
\alias{+.trellis}
\title{lattice plus manipulation}
\usage{
\method{+}{trellis}(p, setting)
}
\arguments{
\item{p}{lattice plot}

\item{setting}{\code{par.settings} for lattice}
}
\description{
lattice plus manipulation
}
\keyword{internal}
