% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{weightedMedian}
\alias{grid.newpage}
\alias{grid.draw}
\alias{grid.rect}
\alias{grid.circle}
\alias{levelplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid]{grid.circle}}, \code{\link[grid]{grid.draw}}, \code{\link[grid]{grid.newpage}}, \code{\link[grid]{grid.rect}}}

  \item{lattice}{\code{\link[lattice]{levelplot}}}

  \item{matrixStats}{\code{\link[matrixStats]{weightedMedian}}}
}}

