% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\alias{get_options}
\title{set latticeGrob options}
\usage{
set_options(options)

get_options(verbose = TRUE)
}
\arguments{
\item{options}{\itemize{
\item style: font style, one of "EN" and "CH"
\item family: font family name, e.g. "Times"
\item family_CH: Chinese fontname for panel.barchartFreq yaxis title
}}
}
\description{
set latticeGrob options
}
