% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lattice.R
\name{set_strip}
\alias{set_strip}
\title{set_strip for lattice}
\usage{
set_strip(p, cex = 1, title = NULL, height = 1)
}
\arguments{
\item{p}{lattice object}
}
\description{
set_strip for lattice
}
\references{
https://stackoverflow.com/questions/17888505/varying-strip-heights-for-multi-panel-lattice-plots
https://stackoverflow.com/questions/8536239/change-background-and-text-of-strips-associated-to-multiple-panels-in-r-lattic
}
