% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_statistic.R
\name{spatial_meansd}
\alias{spatial_meansd}
\title{Statistic of median±sd or median}
\usage{
spatial_meansd(
  z,
  area = NULL,
  digit = 2,
  include.sd = TRUE,
  name = "u",
  unit = "",
  FUN = weightedMedian
)
}
\description{
Statistic of median±sd or median
}
