% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spplot_grid.R
\name{spplot_grid}
\alias{spplot_grid}
\title{Plot methods for spatial data with attributes}
\usage{
spplot_grid(
  grid,
  zcols,
  brks,
  colors,
  strip = FALSE,
  strip.factors = NULL,
  toFactor = FALSE,
  panel.titles_full = NULL,
  panel = panel.spatial,
  xlim = NULL,
  ylim = NULL,
  unit = "",
  unit.adj = 0.3,
  pars = NULL,
  stat = list(show = FALSE, name = "RC", loc = c(81.5, 26.5), digit = 1, include.sd =
    FALSE, FUN = weightedMedian),
  area.weighted = FALSE,
  legend.space = "right",
  legend.num2factor = FALSE,
  colorkey = TRUE,
  interpolate = FALSE,
  lgd.title = NULL,
  sp.layout = NULL,
  layout = NULL,
  NO_begin = 1,
  cex.lgd = 1.3,
  par.settings = opt_trellis_default,
  par.settings2 = list(axis.line = list(col = "transparent")),
  ...
)
}
\arguments{
\item{brks}{either a numeric vector of two or more unique cut points to convert
values in df into factor.}

\item{colors}{corresponding colors of every break interval}

\item{strip}{
    These arguments are described in the help page for
    \code{\link[lattice]{xyplot}}.
  }

\item{panel.titles_full}{string or expression vectors of panel titles, default NULL.
If provided, group names in \code{df} will be overwritten.}

\item{panel}{
    panel function used to create the display, as described in
    \code{\link[lattice]{xyplot}}
  }

\item{xlim}{The limits of x and y}

\item{ylim}{The limits of x and y}

\item{pars}{parameters controlling the position of histogram and panel titles,
e.g. \code{list(title = list(x=77, y=39, cex=1.5), hist = list(origin.x=77, origin.y=28))}.}

\item{colorkey}{Boolean or list returned by \code{\link[=get_colorkey]{get_colorkey()}}.}

\item{sp.layout}{NULL or list; see notes below}

\item{NO_begin}{beginning NO of the first panel}

\item{...}{other parameters to spplot, for example:
\itemize{
\item xlim
\item ylim
\item ...
}}
}
\description{
Lattice (trellis) plot methods for spatial data with attributes
}
\examples{
data("grid_avhrr")

{
    pars = list(title = list(x=77, y=39, cex=1.5), 
                hist = list(origin.x=77, origin.y=28, A=15, by = 0.4))
    stat = list(show = TRUE, name="mean", loc = c(82.5, 26.5), digit = 1, include.sd = TRUE)
    
    p <- spplot_grid(grid_avhrr, 
                     stat = stat, 
                     pars = NULL, 
                     sub.hist = FALSE,
                     interpolate = FALSE)
    p2 <- levelplot2(X1982 ~ s1+s2, grid_avhrr@data, grid_avhrr)
    write_fig(p , "ex-spplot_grid.png", 9.8, 5)
    write_fig(p2, "ex-spplot_grid2.png", 9.8, 5)
}
}
\seealso{
\code{\link[sp:spplot]{sp::spplot()}}, \code{\link[lattice:levelplot]{lattice::levelplot()}}
}
