% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lattice.R
\name{theme_lattice}
\alias{theme_lattice}
\title{theme_lattice}
\usage{
theme_lattice(
  plot.margin,
  key.margin,
  axis.margin,
  axis.components.inner,
  axis.components.outer,
  ...,
  layout.widths,
  layout.heights,
  font_size = NULL,
  font_family = "",
  par.settings
)
}
\arguments{
\item{plot.margin}{plot padding in layout.heights and layoutheights}

\item{key.margin}{legend margin}

\item{axis.margin}{axis margin}

\item{axis.components.inner, axis.components.outer}{margins}

\item{...}{ignored}

\item{layout.widths, layout.heights}{margins of \code{layout.widths} and \code{layout.heights}}

\item{par.settings}{\code{par.settings} for lattice}
}
\description{
margin object in the order to \verb{"top", "right", "bottom", "left"}.
}
\note{
key.margin left not work (right controls)
}
