% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_latFreq.R
\name{layer_latFreq}
\alias{layer_latFreq}
\alias{panel.latFreq}
\title{layer of latitude frequency barchart}
\usage{
layer_latFreq(
  bbox = c(0.7, 1, 0, 1),
  length.out = 10000,
  tcl = 0.4,
  xlab = "",
  ylab = "",
  zlim = NULL,
  zlim_ratio = c(-1, 1),
  prob_z = 0.9,
  is_spatial = FALSE,
  zticks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  digit = 1,
  ...
)

panel.latFreq(
  x,
  y,
  z,
  subscripts,
  bbox = c(0.7, 1, 0, 1),
  length.out = 10000,
  tcl = 0.4,
  xlab = "",
  ylab = "",
  zlim = NULL,
  zlim_ratio = c(-1, 1),
  prob_z = 0.9,
  is_spatial = FALSE,
  zticks = NULL,
  xlabels = NULL,
  ylabels = NULL,
  digit = 1,
  ...
)
}
\arguments{
\item{bbox}{The region to plot \code{grob}, \verb{[xmin, xmax, ymin, ymax]} in the unit
of npc.}

\item{tcl}{The length of tick marks as a fraction of the smaller of the width
or height of the plotting region. If tck >= 0.5 it is interpreted as a fraction
of the relevant side, so if tck = 1 grid lines are drawn. The default setting
(tck = NA) is to use tcl = -0.5.}

\item{xlab, ylab}{the title of xaxis and yaxis.}

\item{zlim}{the limits of \code{z} value. If not specified, it's \code{c(-1, 1)*quantile(abs(z), 0.9)}.}

\item{zlim_ratio}{If \code{zlim} not provided, \code{zlim = zlim_ratio * zmax}}

\item{prob_z}{default 0.9, the probability of z quantile, which used to determine the zlim.}

\item{...}{other parameters to \code{\link[=panel.latFreq]{panel.latFreq()}}}

\item{x}{ Variables defining the plot. }

\item{y}{ Variables defining the plot. }

\item{z}{ Variables defining the plot. }

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }
}
\description{
layer of latitude frequency barchart
}
\examples{
\dontrun{
panel.horizontalFreq(x, y, z, subscripts, bbox = c(0.7, 1, 0, 1), unit = "npc")
}
}
