% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchartFreq.R
\name{panel.barchartFreq}
\alias{panel.barchartFreq}
\title{Panel of frequency barchart}
\usage{
panel.barchartFreq(
  z,
  subscripts,
  bbox = c(0.05, 0.5, 0, 0.5),
  at,
  col.regions,
  box.width = 0.8,
  border = "transparent",
  yscale = NULL,
  yticks = NULL,
  w = NULL,
  ...
)
}
\arguments{
\item{bbox}{The region to plot \code{grob}, \verb{[xmin, xmax, ymin, ymax]} in the unit
of npc.}

\item{at}{
    A numeric vector giving breakpoints along the range of
    \code{z}. Contours (if any) will be drawn at these heights, and the
    regions in between would be colored using \code{col.regions}.  In
    the latter case, values outside the range of \code{at} will not be
    drawn at all.  This serves as a way to limit the range of the data
    shown, similar to what a \code{zlim} argument might have been used
    for.  However, this also means that when supplying \code{at}
    explicitly, one has to be careful to include values outside the
    range of \code{z} to ensure that all the data are shown.

    \code{at} can have length one only if \code{region=FALSE}.  
  }

\item{col.regions}{ color vector to be used if regions is TRUE. The
    general idea is that this should be a color vector of moderately
    large length (longer than the number of regions. By default this is
    100). It is expected that this vector would be gradually varying in
    color (so that nearby colors would be similar). When the colors are
    actually chosen, they are chosen to be equally spaced along this
    vector.  When there are more regions than colors in
    \code{col.regions}, the colors are recycled.  The actual color
    assignment is performed by \code{\link[lattice]{level.colors}}, which is
    documented separately.
  }

\item{box.width}{ Thickness of bars in absolute units; overrides
    \code{box.ratio}.  Useful for specifying thickness when the
    categorical variable is not a factor, as use of \code{box.ratio}
    alone cannot achieve a thickness greater than 1. }

\item{border}{
    Graphical parameters for the bars. By default, the trellis parameter
    \code{plot.polygon} is used if there is no grouping variable, otherwise
    \code{superpose.polygon} is used.  \code{col} gives the fill color,
    \code{border} the border color, and \code{lty} and \code{lwd} the
    line type and width of the borders.
  }

\item{yscale}{frequency range, \verb{[0, 100]}}

\item{yticks}{yaxis ticks position}

\item{...}{
    Further arguments may be supplied.  Some are processed by
    \code{levelplot} or \code{contourplot}, and those that are
    unrecognized are passed on to the panel function.
  }
}
\description{
Panel of frequency barchart
}
