% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.horizontalFreq.R
\name{panel.horizontalFreq}
\alias{panel.horizontalFreq}
\title{Panel of horizontal and vertical polygons}
\usage{
panel.horizontalFreq(
  x,
  y,
  z,
  subscripts,
  bbox = c(0.7, 1, 0, 1),
  unit = "npc",
  length.out = 10000,
  ...,
  tcl = 0.4,
  xlab = "",
  ylab = "",
  zlim = NULL,
  zlim_ratio = c(-1, 1),
  zticks = NULL,
  is_spatial = FALSE,
  prob_z = 0.9,
  xlabels = NULL,
  ylabels = NULL,
  digit = 1
)
}
\arguments{
\item{x}{ Variables defining the plot. }

\item{y}{ Variables defining the plot. }

\item{z}{ Variables defining the plot. }

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }

\item{bbox}{The region to plot \code{grob}, \verb{[xmin, xmax, ymin, ymax]} in the unit
of npc.}

\item{unit}{A character vector specifying the units for the corresponding numeric values.}

\item{...}{ Extra parameters. }

\item{tcl}{The length of tick marks as a fraction of the smaller of the width
or height of the plotting region. If tck >= 0.5 it is interpreted as a fraction
of the relevant side, so if tck = 1 grid lines are drawn. The default setting
(tck = NA) is to use tcl = -0.5.}

\item{xlab, ylab}{the title of xaxis and yaxis.}

\item{zlim}{the limits of \code{z} value. If not specified, it's \code{c(-1, 1)*quantile(abs(z), 0.9)}.}

\item{zlim_ratio}{If \code{zlim} not provided, \code{zlim = zlim_ratio * zmax}}

\item{prob_z}{default 0.9, the probability of z quantile, which used to determine the zlim.}
}
\description{
Panel of horizontal and vertical polygons
}
\examples{
\dontrun{
panel.horizontalFreq(x, y, z, subscripts, bbox = c(0.7, 1, 0, 1), unit = "npc")
}
}
