% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_signPerc.R
\name{panel.signPerc}
\alias{panel.signPerc}
\title{Panel of percentage of negative (significant negative\%) and positive\% (significant positive\%)}
\usage{
panel.signPerc(
  z = NULL,
  mask = NULL,
  col.regions = c("blue", "red"),
  x = 0.05,
  y = 0.8,
  bg = "transparent"
)
}
\arguments{
\item{z}{numeric vector}

\item{mask}{boolean vector with the same length as z, indicating whether
corresponding z value is significant.}

\item{x, y}{The x and y position of positive and negative percentage label.}

\item{bg}{background color}

\item{...}{ignored}
}
\description{
Panel of percentage of negative (significant negative\%) and positive\% (significant positive\%)
}
\examples{
\dontrun{
panel.signPerc(z = NULL, mask = NULL, x = 0.1, y = 0.9, ...)
}
}
