% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.signPerc.R
\name{panel.signPerc}
\alias{panel.signPerc}
\alias{panel.signDist}
\title{Panel of percentage of negative (significant negative\%) and positive\% (significant positive\%)}
\usage{
panel.signPerc(
  z = NULL,
  subscripts,
  mask = NULL,
  xpos = 0.1,
  ypos = 0.9,
  col.regions = c("blue", "red"),
  ...
)

panel.signDist(
  list.mask,
  SpatialPixel,
  par.shade = NULL,
  density = 1,
  angle = 45,
  ...
)
}
\arguments{
\item{z}{numeric vector}

\item{subscripts}{
    Integer vector indicating what subset of \code{x}, \code{y} and
    \code{z} to draw.
  }

\item{mask}{boolean vector with the same length as z, indicating whether
corresponding z value is significant.}

\item{xpos, ypos}{The x and y position of positive and negative percentage label.}

\item{...}{ignored}

\item{list.mask}{NULL or list of masks. \code{mask} is a boolean vector, \code{TRUE}
indicates significant.}

\item{SpatialPixel}{corresponding SpatialPixel object of \code{df}

If \code{mask} present in \code{df}, \code{df.mask} will be ignored.}
}
\description{
Panel of percentage of negative (significant negative\%) and positive\% (significant positive\%)
}
\examples{
\dontrun{
panel.signPerc(z = NULL, mask = NULL, xpos = 0.1, ypos = 0.9, ...)
}
}
