% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sp2.R
\name{sp_area}
\alias{sp_area}
\title{Calculate area of spatial object}
\usage{
sp_area(grid, weighted = TRUE)
}
\arguments{
\item{grid}{SpatialPolygonsDataFrame or SpatialGridDataFrame}

\item{weighted}{if not, ones vector will be return}
}
\description{
Calculate area of spatial object
}
\seealso{
\code{\link[raster:area]{raster::area()}}
}
